% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTAShiny.R
\name{CTAShiny}
\alias{CTAShiny}
\title{Launch CTAShiny Interface}
\usage{
CTAShiny()
}
\value{
Nothing
}
\description{
CTAShiny() loads interactive user interface built using R shiny.
}
\details{
Start CTAShiny

The interactive user interface is to provide an easy way for people who are learning Contingency Table Analysis. Includes example data for testing out a few example analysis.
}
\examples{
if(interactive()){
library(shiny)
CTAShiny()
}
}
\keyword{CTAShiny}
